Added files via upload
I figured out the set-up, but the final draft will be in a different repo for Day 29 HackerRank Challenge. I need to use recursion to create an array with a large enough values to store the output on the bitwise& operator. (Completed). Next I need to go through each value in the list and turn it 0 if the value is greater than another given value. I then need to return the max value in the array list.